refactor: consolidate get_or_create_status into crud_utils
- Removed the duplicate get_or_create_status function from status.py in favor of the existing implementation in crud_utils.py
- Updated all call sites to use explicit keyword arguments (e.g., organization_id=...) to account for the optional description parameter in the crud_utils signature
